Simplify CI branch protection
Configuring GitHub to have many required checks is hard to maintain.

In fact, the admin needs to duplicate what is defined in the actions.yml specs
matrix. It should be quite a bit easier this way, just having one CI job that
depends on the success of the entire specs matrix.